Add Makefile.j2 and Makefile.json
This should make the Makefile a little easier to grok since we can get
rid of the Makefile macros.

In the future we'd be able to get rid of any kind of wildcard rules in
favor of explicitly defining them in Make by using jinja. The benefit of
doing this would be we would have an inspectable file with all
interpolation done vs make doing some internally and it being opaque to
us.

This in turn would pave the way to move to a better build system, like
ninja. Not only is ninja faster than make, it also gives `dot` graphs.