Generate Makefile from Makefile.j2
This gets rid of Make macros which were probably the biggest factor
making thie Makefile hard to understand.

Having jinja2 generate the Makefile can lead to removing any kind of
pattern or wildcard rules, thereby removing the need to parse the
Makefile by acting the part of make. It gets hard to keep all of the
pattern rules and definitions in context while trying to debug something
once a Makefile has grown this large.

I would like to make this Makefile WYSIWYG. At some point in the future
we could convert to Ninja or another tool and take advantage of
performance increases and graphical output of build graph.