API provisioning log event improvements
A few minor tweaks to when certain phone_home log event messages are
sent, and the addition of a new event to declare when the OS image has
been retrieved. The code comments have been updated to reflect the
current state of the event messages we're using for improved
readability.

Signed-off-by: Scott Garman <sgarman@packet.com>