init.next: Implement packet_bootdev_mac handling via new BOOTIF feature
mkinitfs has gained similar functionality as originally implemented with
packet_bootdev_mac so lets just use that and minimize the patch diff.

mkinitfs also learned to not blindly use first eth device. The
implementation is different that what we currently do, but seems to be
fine. It doesn't really matter though since we always set
packet_bootdev_mac when booting with boots which alpine always honors.

Signed-off-by: Manuel Mendez <manny@packet.com>