rules: Do not check binfmt_misc in drone
We don't mount the binfmt_misc psudeo-fs in drone so the check always
fails. We could either mount it when needed or just mess with the
rules so it skips it when built in drone. I chose the latter because it
seems to make more sense to have that handled where the deps are defined
in the first place.

Signed-off-by: Manuel Mendez <manny@packet.com>