Add a way to avoid building Alpine boot files
Since `git` updates the file modified time on clone `make` will want to
rebuild the assets-*/* files. We want to have a way around that in both
ci and for people who may not want to build the files.

So we provide a nice and easy way to skip the building, along with info
in the README and even a very verbose and hopefully eye-catching message
from the Makefile.

Signed-off-by: Manuel Mendez <manny@packet.com>