handlers: Return False from wants_custom_osie if userdata is None
We were only handling the case where `userdata` is not present in the
`instance` map, but cacher data does always has a userdata key set. If
userdata was not supplied for the provision its set to None, which
is returned by `instance.get`.

So we handle userdata being falsy and also take care of empty string.

Signed-off-by: Manuel Mendez <manny@packet.com>