osie-runner: Use pip directly from upstream instead of packaged by Alpine
py3-pip brings in dependencies that conflict with our own, and don't
seem to actually be needed. If those dependencies are in the system,
py3-pip will be more than happy to use them. Later when we delete the
build-deps virtual package those packages go away too and now we can't
start up.

By using a non-package pip we not only avoid competing package managers
, but also don't come across issues with dependencies as pip bundles its
dependencies.

Signed-off-by: Manuel Mendez <manny@packet.com>