allow specific syslog server
Currently the syslog server is pointed to the same tinkerbell host
as used for multiple other requests. This change allows the syslog
server to be defined separately but will fallback to the tinkerbell
host if one is not provided.

To define this simply add `syslog_host=my.host` to the kernel parameters
when booting osie.

Signed-off-by: Mike Mason <mason@packet.com>