Pass git version and drone build information to OSIE containers
This sets some variables in the Makefile to track the OSIE git version
sha, branch, and Drone build number at build time, and passes them along
to the OSIE and osie-runner containers using ARG and ENV docker
variables. These variables are then available to the OSIE scripts and
can be referenced at will.

Signed-off-by: Scott Garman <sgarman@packet.com>