Refactor autofail feature to include tracking timeouts
The autofail() trap function in OSIE now tracks failures in two
contexts:

1. Non-zero exit codes will trigger autofail() and report an error
during the particular OSIE code stage
2. The current OSIE code stage is tracked in /statedir/autofail_stage so
it can bubble up to osie-installer, so we can report where
hangs/timeouts occur.

The $autofail_reason has been renamed to $autofail_stage, which is a
more accurate reflection of what it's used for.

Signed-off-by: Scott Garman <sgarman@packet.com>