Add Discovery OS mode
Discovery OS is a new mode of OSIE that offers various tools to
auto-detect and setup hardware during initial rack enrollment of a
server.

It's based on the packet/discover-metal container image we have
hosted on quay.io. During the OSIE build process we download this image
so it can be available during OSIE runs without having to retrieve it
via quay.

Signed-off-by: Scott Garman <sgarman@packet.com>