Download our BIOS config files during deprov
The bios-configs-latest tarball includes our BIOS config files and a
manifest.txt file which provides a lookup table to map a plan name and
BIOS vendor to a configuration filename. The manifest also has an
enforcement status field which can be set to "testing" if we want to
only check and report BIOS drift but not apply the config.

Signed-off-by: Scott Garman <sgarman@packet.com>