Detect and report BIOS configuration drift on deprov
These functions compare the current BIOS configuration to what we're
expecting from the configs in our firmware repo. This isn't as simple as
dumping the config and performing a diff - the configs must be
normalized by filtering out irrelevant config sections that might change
from server to server, but don't represent meaningful drift.

The filtering approach used here is admittedly messy and not very
maintainable over the long term, but this effort is being done as a
stopgap until a separate, microservice-based approach to maintaining
firmware and BIOS settings is completed.

Signed-off-by: Scott Garman <sgarman@packet.com>