ci: build a custom proxy binary using examples from caddyserver and http.upload
mholt/caddy repository has gone private, so we need to use the
official caddy repo instead and use a newer version that doesn't include
go dependencies on the old, private repo.

I found it easier to use go.mod to manage dependencies vs downloading
from GitHub in the Dockerfile. This will likely not work if go is
updated due to a runtime issue with quic-go (unless we fork something).
At that point we might as well either use a "real" tinkerbell stack or
switch to HAProxy over caddy.

Fixes #176
Closes #175

Signed-off-by: Scott Garman <sgarman@packet.com>
Signed-off-by: Manuel Mendez <mmendez@equinix.com>