Check enforcement status of BIOS config before applying
In the bios configs tarball we have a manifest.txt file which includes
an enforcement status field. If this field is not set to "enforce", all
BIOS config checks will happen, but the config will not be applied. This
allows us to improve the config normalization filters during a
development period, after which we can set the enforcement status to
enable the config to be written when drift is detected.

Signed-off-by: Scott Garman <sgarman@packet.com>