Do not fail when SOL causes docker to return error
If SOL isn't working properly, docker can return an error even though
the script has run to successful completion. This change sets the
autofail message to "completed" at the very end of the (de)provision
scripts so that if docker errors we can detect that script completed
and cleanup gracefully instead of failing.

Signed-off-by: Scott Garman <sgarman@packet.com>
Signed-off-by: Manuel Mendez <mmendez@equinix.com>