Do not fail deprovisions when SOL causes docker to hang
If SOL isn't working properly during deprovision, docker can hang even
though the deprovision script has run to completion. This change sets
the autofail message to "deprovision completed" at the very end of the
deprovision script, so that if docker hangs and is killed by the timeout
watcher, we can detect that the deprovision script completed and cleanup
gracefully instead of failing the deprovision.

Signed-off-by: Scott Garman <sgarman@packet.com>