Improve provisioning logging for Windows and FreeBSD OSes
This adds the autofail() bash trap function to these installers in the
same way we use it in the osie.sh and deprovision.sh scripts. Generates
logging error messages at various markers through the provisioning
process, so that if the script fails we can report the last stage
reached instead of a generic error/timeout message.

Signed-off-by: Scott Garman <sgarman@packet.com>