wosie.sh: remove ping-based check for OS image source
In osie.sh, we replaced the use of ensure_reachable() with the
github_mirror_check() function, to avoid the use of ping, which doesn't
work as a valid test in our k8s facilities. In wosie.sh, we test that
the OS image url is accessible with wget, so ensure_reachable() isn't
needed anymore.

Signed-off-by: Scott Garman <sgarman@packet.com>