Route new github LFS hostname via IPN
GitHub appears to have changed their LFS backend to use a new hostname,
github-cloud.githubusercontent.com. In order to route this via IPN for
caching purposes, add it to the /etc/hosts file using the anycast IP for
IPN.

It's not yet confirmed whether they've stopped using the AWS s3 hostname,
so we're leaving that in place for the time being.

Signed-off-by: Scott Garman <sgarman@equinix.com>