make: Set .ONESHELL and update rules.mk.j2
This has the side effect of making $(Q) useless, so get rid of it.

Q is useless because we won't see the commands run when V!=1 due to
`E = @echo`. Since make is now in ONESHELL mode E's @ applies to the
entire recipe. This is actually an unexpected benefit IMO :D .

Signed-off-by: Manuel Mendez <mmendez@equinix.com>