Avoid sending empty BIOS invetories to Hollow
The 'packet-hardware inventorybios' command can succeed but generate
empty BIOS JSON on unsupported platforms. Check for this and avoid
writing to Hollow in these cases.

The $id that we get from metadata is the instance uuid, not the
hardware uuid, which we need to use for Hollow data. Make the hwuuid
available to osie scripts by passing it down via osie runner's docker
env.

Signed-off-by: Scott Garman <sgarman@equinix.com>