generate a traceparent if none was found, return it
I was plumbing traceparent down to the docker commands and started
adding `if traceparent is None` checks and realized that's not great to
copy around and it would be better if traceparent is always populated
and safe to use, so this does that.

When no traceparent is in kernel command line or in the environment, it
will start a fresh unparented span and return it as the traceparent.

Signed-off-by: Amy Tobey <atobey@equinix.com>