plumb otel traceparent through to docker run
Adds traceparent plumbing to Handler, adding it as a member and
appending it to the environments passed to `run_osie` at each callsite.
I did it this way because for some reason `run_osie` is a static method
and does not have a `self` so this gets everything we want well enough
without changing that.